+++ ScriptureRenderingPipeline/RenderingTrigger.cs
@@ -18,7 +18,6 @@ using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using PipelineCommon.Helpers;
 using PipelineCommon.Models.BusMessages;
-using PipelineCommon.Models.ResourceContainer;
 using ScriptureRenderingPipeline.Models;
 using ScriptureRenderingPipeline.Renderers;
 using YamlDotNet.Serialization;
@@ -314,21 +313,3 @@ public static class RenderingTrigger
 		return await new StreamReader(templateStream).ReadToEndAsync();
 	}
 }
-    public class RendererInput
-    {
-	   public string UserToRouteResourcesTo { get; set; } 
-	   public string BaseUrl { get; set; }
-	   public Template PrintTemplate { get; set; }
-	   public ZipFileSystem FileSystem { get; set; }
-	   public string BasePath { get; set; }
-	   public bool IsBTTWriterProject { get; set; }
-	   public string ResourceName { get; set; }
-	   public string LanguageName { get; set; }
-	   public string LanguageCode { get; set; }
-	   public string LanguageTextDirection { get; set; }
-	   public AppMeta AppsMeta { get; set; }
-	   public ResourceContainer ResourceContainer { get; set; }
-	   public string OutputDir { get; set; }
-	   public string Title { get; set; }
-	   public string RepoUrl { get; set; }
-    }
